Move categorization to core package
